:ambulance: dell-smm-hwmon: Don't depend on kernel
By forcing the kernel dependency to the exact version that
dell-smm-hwmon was build with, pacman can no longer upgrade past this
kernel version. This is undesirable and breaks system upgrades until
dell-smm-hwmon gets rebuilt.

We shouldn't get in the way of upgrades, so instead the depends is
removed and we install a pacman/ALPM hook to be reminded of the need to
rebuild dell-smm-hwmon once linux or linux-lts gets installed/updated.