desktop: Add code
This brings back VS Code, but now using the open source build from the
community repo. This is enough for our needs since we need none of the
proprietary "have to be Microsoft" extensions from their marketplace and
can instead rely on the open source build and the OpenVSX marketplace.